Move groups sub-apis to separate folder
